refactor: replace deprecated constants file with app_constants and update imports across the project
