--- lib/core/constants/constant.dart
@@ -1,23 +1,0 @@
-// lib/constants/constants.dart
-import 'package:flutter/material.dart';
-
-class AppSpacing {
-  static const double horizontal = 16.0;
-  static const double vertical = 8.0;
-  static const double cardSpacing = 8.0;
-}
-
-class AppBarInfo {
-  static const double height = 80.0;
-  static const double elevation = 1.0;
-  static const Color backgroundColor = Colors.white;
-  static const Color shadowColor = Colors.grey;
-  static const TextStyle titleTextStyle = TextStyle(
-    fontSize: 18,
-    fontWeight: FontWeight.w500,
-  );
-  static const TextStyle subtitleTextStyle = TextStyle(
-    fontSize: 14,
-    fontWeight: FontWeight.w400,
-  );
-}
