+++ lib/presentation/views/bots/bots_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:khtn_ai_final_project/core/constants/constants.dart';
+import 'package:khtn_ai_final_project/core/constants/app_constants.dart';
 import 'package:khtn_ai_final_project/presentation/common/widgets/bot_search_bar.dart';
 import 'package:khtn_ai_final_project/presentation/viewmodels/bot_view_model.dart';
 import 'package:khtn_ai_final_project/core/theme/app_radius.dart';
