+++ lib/presentation/views/chat/widgets/chat_app_bar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:khtn_ai_final_project/core/constants/constants.dart';
+import 'package:khtn_ai_final_project/core/constants/app_constants.dart';
 import 'bot_option_menu.dart';
 
 class ChatAppBar extends StatelessWidget implements PreferredSizeWidget {
